requirements.txt
scikit-learn
pandas
numpy
pusher
joblib
python-dotenv
websocket-client
twilio